Minor UI changes on the navigation bar. Reorganised folder structure. Decomposed Favours page into smaller components.
