+++ .gitignore
@@ -1,2 +1,3 @@
 /client/node_modules
 node_modules/
+.idea
