--- client/src/Components/Pages/favours/Favours.css
@@ -1,264 +1,0 @@
-#Favours {
-    display: flex;
-    flex-direction: row;
-    padding-top: 80px;
-    justify-content: space-between;
-    height: calc(100vh - 82px);
-    min-width: 100%;
-    background-color: #f3f0ea;
-    margin-left: 0;
-    transition: 2s;
-}
-.all_favours_area {
-    display: flex;
-    flex-direction: column;
-    min-width: 800px;
-    width: 50%;
-    align-items: flex-start;
-    padding: 0 40px 0 20px;
-    margin-left: 6%;
-    margin-right: 20px;
-}
-.favours_title {
-    font-weight: 600;
-    color: #000000;
-    font-size: 60px;
-    font-family: "Poppins";
-    padding-left: 15px;
-    margin-top: 30px;
-}
-.filter_btn {
-    display: flex;
-    flex-direction: row;
-    min-width: 300px;
-    text-align: justify;
-    font-size: 50px;
-    padding-bottom: 15px;
-    outline: none;
-    padding-left: 15px;
-}
-
-.all_btn, .owe_btn, .owing_btn, .settled_btn {
-    height: 30px;
-    font-size: 18px;
-    font-weight: 500;
-    font-family: Roboto;
-    background-color: transparent;
-    color: black;
-}
-button {
-    display: flex;
-    flex: auto;
-    border: medium none;
-    color: #000000;
-    font-size: 16px;
-    background: #fff;
-}
-button {
-    color: #000;
-    transition: all 200ms ease-in-out;
-}
-button:hover {
-    opacity: 0.7;
-    color: dodgerblue;
-}
-button:focus {
-    font-weight: bold;
-    color: dodgerblue;
-    outline:0;
-}
-.favours_cards {
-    display: block;
-    min-width: 400px;
-    padding: 10px 10px 10px 10px;
-    overflow: auto;
-}
-.favour_card_left {
-    display: flex;
-    flex-direction: column;
-    padding: 10px 20px 10px 20px;
-    width: 420px;
-    min-height: 65px;
-    transition: 0.2s;
-    margin-bottom: 20px;
-    background-image: linear-gradient(to bottom left, #e5e5e5, #dedede);
-    text-align: left;
-    justify-content: space-evenly;
-    border-radius: 15px;
-    box-shadow: 1px 1px 7px rgb(227, 227, 227);
-}
-.favour_card_right {
-    display: flex;
-    flex-direction: column;
-    padding: 10px 20px 10px 20px;
-    width: 420px;
-    min-height: 65px;
-    transition: 0.2s;
-    margin-bottom: 20px;
-    margin-left: 22vw;
-    background-image: linear-gradient(to bottom left, #7e51d0, #4b83e9);
-    text-align: right;
-    justify-content: space-evenly;
-    align-content: flex-end;
-    border-radius: 15px;
-    box-shadow: 1px 1px 7px rgb(145, 145, 145);
-}
-
-.favour_card_right:hover, .favour_card_left:hover {
-    box-shadow: 2px 2px 10px rgb(153, 153, 153);
-}
-.user_label_left{
-    font-family: "Poppins";
-    font-weight: bold;
-    font-size: 19px;
-    color: black;
-}
-.card_content_left {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-}
-.value_label_left{
-    font-family: "Roboto";
-    font-size: 16px;
-    color: black;
-}
-.date_left {
-    font-family: "Roboto";
-    font-size: 16px;
-    font-weight: 500;
-    color: #515151;
-}
-.user_label_right{
-    font-family: "Poppins";
-    font-weight: bold;
-    font-size: 19px;
-    color: white;
-}
-.value_label_right{
-    font-family: "Roboto";
-    font-size: 17px;
-    color: white;
-}
-.card_content_right {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-}
-.date_right {
-    font-family: "Roboto";
-    font-size: 16px;
-    font-weight: 500;
-    color: #ffffff;
-}
-
-.create_favours_area {
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-    justify-content: center;
-    padding: 0 30px 0 30px;
-    margin-right: 150px;
-    margin-top: 30px;
-    min-width: 10vw;
-}
-.create_favours_card {
-    display: flex;
-    flex-direction: column;
-    min-width: 25vw;
-    min-height: fit-content;
-    height: 75%;
-    justify-content: center;
-    align-items: center;
-    box-shadow: 0 0 3px 3px rgba(133, 135, 141, 0.12);
-    padding-left: 6%;
-    padding-right: 6%;
-    border-radius: 12px;
-}
-.create_favours_card:hover {
-    /* for right side the, -shadow: -2px -2px 15px 0 rgba(0,0,0,0.2); */
-    box-shadow: -2px -2px 10px 0 rgb(226, 226, 226);
-}
-.create_favour_content {
-    width: fit-content;
-}
-.create_favour_label {
-
-}
-.create_favour_form {
-    height: fit-content;
-    width: fit-content;
-    padding-bottom: 20px;
-}
-.btn_container {
-    display: flex;
-    flex-direction: row;;
-    height: 40px;
-    width: fit-content;
-    justify-content: flex-start;
-    align-items: flex-start;
-}
-.create_favour_btn {
-    width: 160px;
-    height: inherit;
-    background-color: hotpink;
-    color: white;
-    font-weight: bold;
-    font-size: 17px;
-    border-radius: 25px;
-    justify-content: center;
-    align-items: center;
-}
-.cancel_favour_btn {
-    background-color: transparent;
-    color: black;
-    size: 30px;
-    vertical-align: middle;
-}
-.create_favour_first_line {
-    display: flex;
-    align-content:flex-start;
-    width: inherit;
-    font-family: Poppins;
-    font-weight: bold;
-    font-size: 40px;
-}
-.create_favour_second_line {
-    display: flex;
-    align-content:flex-start;
-    font-family: Poppins;
-    width: inherit;
-    font-weight: bold;
-    font-size: 40px;
-}
-.create_favour_third_line{
-    display: flex;
-    flex-direction: row;
-    font-family: Poppins;
-    font-weight: bold;
-    font-size: 40px;
-}
-.favour_type_field {
-    color: hotpink;
-}
-.lets_label, .a_label, .from_label {
-    font-family: Poppins;
-    font-weight: bold;
-    font-size: 40px;
-    padding-right: 9px;
-}
-.value_field {
-    color: hotpink;
-    padding-right: 9px;
-}
-.who_field {
-    display: flex;
-    flex-direction: row;
-    width: 500px;
-}
-
-
-.heidi {
-    cheek-width: smol;
-    waist-width: smol;
-}
