+++ client/src/Components/favour/Favour.js
@@ -1,0 +1,129 @@
+import React from 'react';
+import "./Favour.css";
+
+const Favour = () => {
+
+    return (
+        <div className="favours_cards">
+            <div className= "favour_card_left">
+                <div className="user_label_left">
+                    @PabloPicasso
+                </div>
+                <div className = "card_content_left">
+                    <div className="value_label_left">
+                        Owes you <strong> coffee </strong>
+                    </div>
+                    <div className="date_left">
+                        30 Aug
+                    </div>
+                </div>
+            </div>
+            <div className= "favour_card_left">
+                <div className="user_label_left">
+                    @MarkRothko
+                </div>
+                <div className = "card_content_left">
+                    <div className="value_label_left">
+                        Owes you <strong> lunch </strong>
+                    </div>
+                    <div className="date_left">
+                        29 Aug
+                    </div>
+                </div>
+            </div>
+            <div className= "favour_card_right">
+                <div className="user_label_right">
+                    @You
+                </div>
+                <div className="card_content_right">
+                    <div className="date_right">
+                        28 Aug
+                    </div>
+                    <div className="value_label_right">
+                        Owe <strong> @EdwardHopper </strong> <strong> lunch </strong>
+                    </div>
+                </div>
+            </div>
+            <div className= "favour_card_left">
+                <div className="user_label_left">
+                    @HenriMatisse
+                </div>
+                <div className = "card_content_left">
+                    <div className="value_label_left">
+                        Owes you <strong> drinks </strong>
+                    </div>
+                    <div className="date_left">
+                        27 Aug
+                    </div>
+                </div>
+            </div>
+            <div className= "favour_card_right">
+                <div className="user_label_right">
+                    @You
+                </div>
+                <div className="card_content_right">
+                    <div className="date_right">
+                        26 Aug
+                    </div>
+                    <div className="value_label_right">
+                        Owe <strong> @MarkRothko </strong> <strong> coffee </strong>
+                    </div>
+                </div>
+            </div>
+            <div className= "favour_card_right">
+                <div className="user_label_right">
+                    @You
+                </div>
+                <div className="card_content_right">
+                    <div className="date_right">
+                        25 Aug
+                    </div>
+                    <div className="value_label_right">
+                        Owe <strong> @HernanBas </strong> <strong> drinks </strong>
+                    </div>
+                </div>
+            </div>
+            <div className= "favour_card_left">
+                <div className="user_label_left">
+                    @PabloPicasso
+                </div>
+                <div className = "card_content_left">
+                    <div className="value_label_left">
+                        Owes you <strong> drinks </strong>
+                    </div>
+                    <div className="date_left">
+                        25 Aug
+                    </div>
+                </div>
+            </div>
+            <div className= "favour_card_left">
+                <div className="user_label_left">
+                    @JacksonPollock
+                </div>
+                <div className = "card_content_left">
+                    <div className="value_label_left">
+                        Owes you <strong> coffee </strong>
+                    </div>
+                    <div className="date_left">
+                        23 Aug
+                    </div>
+                </div>
+            </div>
+            <div className= "favour_card_right">
+                <div className="user_label_right">
+                    @You
+                </div>
+                <div className="card_content_right">
+                    <div className="date_right">
+                        20 Aug
+                    </div>
+                    <div className="value_label_right">
+                        Owe <strong> @JeanMichel </strong> <strong> coffee </strong>
+                    </div>
+                </div>
+            </div>
+        </div>
+    )
+}
+
+export default Favour;
