--- client/src/Components/navbar/Navbar.css
@@ -1,74 +1,0 @@
-
-#navbar {
-    min-width: 100%;
-    height: 85px;
-    box-shadow: 0 0.5rem 0.625rem rgba(41, 42, 45, 0.12);
-    display: flex;
-    color: #000;
-    background-color: #f3f0ea;
-    position: absolute;
-    justify-content: space-between;
-}
-#navbar > div {
-    height: 100%;
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    cursor: pointer;
-}
-.logo {
-    min-width: fit-content;
-    font-size:25px;
-    font-weight: bold;
-    color: black;
-    font-family: Poppins;
-}
-.page_links {
-    display: flex;
-    font-size: 15px;
-    margin-left: 25px;
-}
-.links_label {
-    padding-left: 18px;
-    padding-right: 18px;
-}
-.links_label a {
-    text-decoration: none;
-    color: black;
-}
-ul {
-    display: flex;
-    margin: 0;
-}
-ul li {
-    padding: 0 16px;
-}
-a {
-    color: #000;
-    transition: all 200ms ease-in-out;
-}
-a:hover {
-    opacity: 0.7;
-    color: dodgerblue;
-}
-a:focus {
-    font-weight: bold;
-}
-.authentication_links {
-    display: flex;
-    font-size: 15px;
-    margin-right: 30px;
-}
-.signup_link_label {
-    width: 120px;
-    height: 35px;
-    background-color: transparent;
-    color: black;
-    border: thin solid black;
-    font-weight: bold;
-    font-size: 15px;
-    border-radius: 20px;
-    margin-left: 10px;
-    display:table-cell;
-    vertical-align:middle;
-}
