+++ client/src/navbar/Navbar.css
@@ -1,0 +1,173 @@
+
+#navbar {
+    width: 100%;
+    height: 85px;
+    box-shadow: 0 0.5rem 0.625rem rgba(41, 42, 45, 0.12);
+    display: flex;
+    color: #000;
+    background-color: black;
+    position: absolute;
+    justify-content: space-between;
+    align-items: center;
+    z-index: 0;
+}
+#navbar > div {
+    height: 100%;
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    cursor: pointer;
+}
+.logo {
+    min-width: fit-content;
+    font-size: 30px;
+    font-weight: bold;
+    color: white;
+    font-family: Poppins;
+}
+.page_links {
+    padding-right: 10px;
+    font-size: 20px;
+    font-weight: 550;
+    text-decoration: none;
+    width: 200px;
+}
+.page_links a {
+    text-decoration: none;
+    color: black;
+}
+a {
+    color: white;
+    transition: all 200ms ease-in-out;
+}
+a:hover {
+    opacity: 0.7;
+    color: #016acf;
+}
+a:focus {
+    font-weight: 900;
+}
+
+ul {
+    display: flex;
+    margin: 0;
+}
+ul li {
+    padding: 0 16px;
+}
+
+.authentication_links {
+    display: flex;
+    font-size: 12px;
+    margin-right: 40px;
+    justify-content: space-between;
+    width: 180px;
+}
+.authentication_links a {
+    text-decoration: none;
+    color: white;
+    font-size: 14px;
+    font-weight: bold;
+}
+.signup_link_label {
+    width: 120px;
+    height: 35px;
+    background-color: transparent;
+    color: white;
+    border: thin solid white;
+    font-weight: bold;
+    font-size: 14px;
+    border-radius: 20px;
+    display:table-cell;
+    vertical-align: middle;
+}
+
+
+
+/* here navbar */
+
+nav {
+    background-color: transparent;
+    height: 65px;
+}
+
+#menuToggle {
+    display: flex;
+    flex-direction: column;
+    position: relative;
+    top: 25px;
+    left: 30px;
+    z-index: 1;
+    -webkit-user-select: none;
+    user-select: none;
+    width: 40px;
+}
+#menuToggle input {
+    display: flex;
+    width: 40px;
+    height: 32px;
+    position: absolute;
+    cursor: pointer;
+    opacity: 0;
+    z-index: 2;
+
+}
+#menuToggle span {
+    display: flex;
+    width: 20px;
+    height: 2px;
+    margin-bottom: 5px;
+    position: relative;
+    background: white;
+    border-radius: 3px;
+    z-index: 1;
+    transform-origin: 5px 0px;
+    transition: transform 0.5s cubic-bezier(0.77,0.2,0.05,1.0),
+    background 0.5s cubic-bezier(0.77,0.2,0.05,1.0),
+    opacity 0.55s ease;
+}
+#menuToggle span:first-child {
+    transform-origin: 0% 0%;
+}
+#menuToggle span:nth-last-child(2) {
+    transform-origin: 0% 100%;
+}
+#menuToggle input:checked ~ span {
+    opacity: 1;
+    transform: rotate(45deg) translate(-3px, -1px);
+}
+#menuToggle input:checked ~ span:nth-last-child(3) {
+    opacity: 0;
+    transform: rotate(0deg) scale(0.2, 0.2);
+}
+#menuToggle input:checked ~ span:nth-last-child(2) {
+    transform: rotate(-45deg) translate(0, -1px);
+}
+#menu {
+    display: flex;
+    flex-direction: column;
+    text-align: left;
+    position: absolute;
+    width: 110px;
+    height: 200px;
+    justify-content: space-evenly;
+    margin: -50px 0 0 -50px;
+    padding: 100px 50px 50px;
+    background-color: transparent;
+    -webkit-font-smoothing: antialiased;
+    transform-origin: 0% 0%;
+    transform: translate(-100%, 0);
+    transition: transform 0.5s cubic-bezier(0.77,0.2,0.05,1.0);
+}
+#menu li {
+    padding: 10px 0;
+    transition-delay: 2s;
+}
+
+#menuToggle input:checked ~ ul {
+    transform: none;
+}
+
+.test {
+    transform: translateX(-45%);
+}
